backend/app/requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
redis>=5.0
django-redis
djangorestframework
djangorestframework-simplejwt
dotenv
cryptography